requirements.txt
Django<4.0
django-debug-toolbar
django-environ
django-extensions
django-ckeditor==6.2.0
Pillow
pytils
django-tagulous
django-versatileimagefield
django-widget-tweaks
django-extra-views
djangorestframework
